Update for time TBA and remove Event crud
This commit updates the display to show concerts that start exactly at
midnight as 'TBA' for concerts where we don't know the exact time.

It also removes CRUD behavior from the routes, Event controller and
views for events, since all editing and creating is handled by the
Active Admin gem.